requirements.txt
streamlit==1.52.2
python-dotenv==1.2.1

langchain==1.2.0
langchain-openai==1.1.6
langchain-community==1.0.0a1
langchain-text-splitters==1.1.0

chromadb==1.4.0
pypdf==6.5.0

trafilatura==2.0.0
beautifulsoup4==4.14.3
requests==2.32.5
